chore: Push `combineContexts` to utilities
