+++ utilities/combine-contexts.js
@@ -1,0 +1,19 @@
+const _ = require('lodash');
+
+function combineContexts(context1, context2) {
+  function webpackContext(req) {
+    try {
+      return context1(req);
+    } catch (e) {
+      return context2(req);
+    }
+  }
+  webpackContext.keys = () => {
+    let keys1 = context1.keys();
+    let keys2 = context2.keys();
+    return _.chain(keys1).concat(keys2).sortBy().uniq().value();
+  };
+  return webpackContext;
+}
+
+module.exports = combineContexts;
